+++ tests/test_download.py
@@ -10,10 +10,20 @@ def test_system_mmsdm(base_dir: pathlib.Path) -> None:
         download(start=month, end=month, table="trading-price", base_directory=base_dir)
         assert (base_dir / "trading-price" / month / "clean.parquet").exists()
 
+    repeat_months = ["2020-02", "2020-02"]
+    for month in repeat_months:
+        download(start=month, end=month, table="trading-price", base_directory=base_dir)
+        assert (base_dir / "trading-price" / month / "clean.parquet").exists()
+
 
 def test_system_nemde(base_dir: pathlib.Path) -> None:
     days = ["2020-01-01", "2020-01-02"]
     for day in days:
         assert not (base_dir / "nemde" / day / "clean.parquet").exists()
+        download(start=day, end=day, table="nemde", base_directory=base_dir)
+        assert (base_dir / "nemde" / day / "clean.parquet").exists()
+
+    repeat_days = ["2020-01-02", "2020-01-02"]
+    for day in repeat_days:
         download(start=day, end=day, table="nemde", base_directory=base_dir)
         assert (base_dir / "nemde" / day / "clean.parquet").exists()